feat: restructure CLI commands by removing deprecated scan module and adding new evaluate and scan commands
